+++ sdk/communication/azure-communication-common/src/main/java/com/azure/communication/common/CommunicationTokenCredential.java
@@ -83,37 +83,30 @@ public final class CommunicationTokenCredential implements AutoCloseable {
     /**
      * Create with a tokenRefresher
      * 
-     * @param tokenRefresher implementation to supply fresh token when reqested
+     * @param tokenRefreshOptions implementation to supply fresh token when reqested
      */
-    public CommunicationTokenCredential(TokenRefresher tokenRefresher) {
+    public CommunicationTokenCredential(TokenRefreshOptions tokenRefreshOptions) {
+        TokenRefresher tokenRefresher = tokenRefreshOptions.getTokenRefresher();
         Objects.requireNonNull(tokenRefresher, "'tokenRefresher' cannot be null.");
         refresher = tokenRefresher;
     }
 
     /**
      * Create with serialized JWT token and a token supplier to auto-refresh the
-     * token before it expires. Callback function tokenRefresher will be called ahead
-     * of the token expiry by the number of minutes specified by
+     * token before it expires. Callback function tokenRefresher will be called
+     * ahead of the token expiry by the number of minutes specified by
      * CallbackOffsetMinutes defaulted to two minutes. To modify this default, call
      * setCallbackOffsetMinutes after construction
      * 
-     * @param tokenRefresher implementation to supply fresh token when reqested
+     * @param tokenRefreshOptions implementation to supply fresh token when reqested
      * @param initialToken serialized JWT token
-     * @param refreshProactively when set to true, turn on proactive fetching to
-     *                           call tokenRefresher before token expiry by minutes
-     *                           set with setCallbackOffsetMinutes or default value
-     *                           of two minutes
      */
-    public CommunicationTokenCredential(
-        TokenRefresher tokenRefresher, 
-        String initialToken,
-        boolean refreshProactively) 
-    {
-        this(tokenRefresher);
+    public CommunicationTokenCredential(TokenRefreshOptions tokenRefreshOptions, String initialToken) {
+        this(tokenRefreshOptions);
         Objects.requireNonNull(initialToken, "'initialToken' cannot be null.");
         setToken(initialToken);
         tokenFuture = new TokenImmediate(accessToken);
-        if (refreshProactively) {
+        if (tokenRefreshOptions.getRefreshProactively()) {
             OffsetDateTime nextFetchTime = accessToken.getExpiresAt().minusMinutes(DEFAULT_EXPIRING_OFFSET_MINUTES);
             fetchingTask = new FetchingTask(this, nextFetchTime);
         }