+++ sdk/communication/azure-communication-common/src/test/java/com/azure/communication/common/CommunicationTokenCredentialTests.java
@@ -112,8 +112,9 @@ public class CommunicationTokenCredentialTests {
     public void fresherShouldNotBeCalledBeforeExpiringTime() throws InterruptedException, ExecutionException, IOException {
         String tokenStr = tokenMocker.generateRawToken("resourceId", "userIdentity", 15 * 60);
         immediateFresher.resetCallCount();
-        CommunicationTokenCredential tokenCredential = new CommunicationTokenCredential(immediateFresher, tokenStr, true);
-        AccessToken token = tokenCredential.getToken().get();
+
+        TokenRefreshOptions tokenRefreshOptions = new TokenRefreshOptions(immediateFresher, true);
+        CommunicationTokenCredential tokenCredential = new CommunicationTokenCredential(tokenRefreshOptions, tokenStr);        AccessToken token = tokenCredential.getToken().get();
         assertFalse(token.isExpired(),
                 "Refreshable AccessToken should not expire when expiry is set to 5 minutes later");
         assertEquals(tokenStr, token.getToken());
@@ -127,8 +128,8 @@ public class CommunicationTokenCredentialTests {
         immediateFresher.resetCallCount();
         CountDownLatch countDownLatch = new CountDownLatch(1);
         immediateFresher.setOnCallReturn(countDownLatch::countDown);
-        CommunicationTokenCredential tokenCredential = new CommunicationTokenCredential(immediateFresher, tokenStr, true);
-        
+        TokenRefreshOptions tokenRefreshOptions = new TokenRefreshOptions(immediateFresher, true);
+        CommunicationTokenCredential tokenCredential = new CommunicationTokenCredential(tokenRefreshOptions, tokenStr);        
         countDownLatch.await();
         assertEquals(1, immediateFresher.numCalls());
         AccessToken token = tokenCredential.getToken().get();
@@ -143,8 +144,8 @@ public class CommunicationTokenCredentialTests {
         immediateFresher.resetCallCount();
         CountDownLatch countDownLatch = new CountDownLatch(1);
         immediateFresher.setOnCallReturn(countDownLatch::countDown);
-        CommunicationTokenCredential tokenCredential = new CommunicationTokenCredential(immediateFresher, tokenStr, true);
-        
+        TokenRefreshOptions tokenRefreshOptions = new TokenRefreshOptions(immediateFresher, true);
+        CommunicationTokenCredential tokenCredential = new CommunicationTokenCredential(tokenRefreshOptions, tokenStr);        
         countDownLatch.await();
         assertEquals(1, immediateFresher.numCalls());
         AccessToken token = tokenCredential.getToken().get();
@@ -158,8 +159,8 @@ public class CommunicationTokenCredentialTests {
         immediateFresher.resetCallCount();
         CountDownLatch firstCountDownLatch = new CountDownLatch(1);
         immediateFresher.setOnCallReturn(firstCountDownLatch::countDown);
-        CommunicationTokenCredential tokenCredential = new CommunicationTokenCredential(immediateFresher, tokenStr, true);
-        
+        TokenRefreshOptions tokenRefreshOptions = new TokenRefreshOptions(immediateFresher, true);
+        CommunicationTokenCredential tokenCredential = new CommunicationTokenCredential(tokenRefreshOptions, tokenStr);        
         firstCountDownLatch.await();
         assertEquals(1, immediateFresher.numCalls());
         AccessToken token = tokenCredential.getToken().get();
@@ -179,7 +180,8 @@ public class CommunicationTokenCredentialTests {
     public void shouldNotCallRefresherWhenTokenStillValid() throws InterruptedException, ExecutionException, IOException {
         String tokenStr = tokenMocker.generateRawToken("resourceId", "userIdentity", 15 * 60);
         immediateFresher.resetCallCount();
-        CommunicationTokenCredential tokenCredential = new CommunicationTokenCredential(immediateFresher, tokenStr, false);
+        TokenRefreshOptions tokenRefreshOptions = new TokenRefreshOptions(immediateFresher, false);
+        CommunicationTokenCredential tokenCredential = new CommunicationTokenCredential(tokenRefreshOptions, tokenStr);
         AccessToken token = tokenCredential.getToken().get();
         assertFalse(token.isExpired());
         assertEquals(tokenStr, token.getToken());
@@ -196,7 +198,8 @@ public class CommunicationTokenCredentialTests {
     public void expiredTokenShouldBeRefreshedOnDemandWithoutProactiveFetch() throws InterruptedException, ExecutionException, IOException {
         String tokenStr = tokenMocker.generateRawToken("resourceId", "userIdentity", -5 * 60);
         immediateFresher.resetCallCount();
-        CommunicationTokenCredential tokenCredential = new CommunicationTokenCredential(immediateFresher, tokenStr, false);
+        TokenRefreshOptions tokenRefreshOptions = new TokenRefreshOptions(immediateFresher, false);
+        CommunicationTokenCredential tokenCredential = new CommunicationTokenCredential(tokenRefreshOptions, tokenStr);
         assertEquals(0, immediateFresher.numCalls());
         AccessToken token = tokenCredential.getToken().get();
         assertFalse(token.isExpired(),
@@ -214,7 +217,8 @@ public class CommunicationTokenCredentialTests {
     @Test
     public void shouldCallbackOnDemandWithoutRefresher() throws InterruptedException, ExecutionException, IOException {
         immediateFresher.resetCallCount();
-        CommunicationTokenCredential tokenCredential = new CommunicationTokenCredential(immediateFresher);
+        TokenRefreshOptions tokenRefreshOptions = new TokenRefreshOptions(immediateFresher, true);
+        CommunicationTokenCredential tokenCredential = new CommunicationTokenCredential(tokenRefreshOptions);
         AccessToken accessToken = tokenCredential.getToken().get();
         assertEquals(1, immediateFresher.numCalls());
         assertFalse(accessToken.isExpired(), "On demand fetching case, should be still valid");
@@ -225,7 +229,8 @@ public class CommunicationTokenCredentialTests {
     @Test
     public void shouldStopRefreshTimerWhenClosed() throws InterruptedException, ExecutionException, IOException {
         String tokenStr = tokenMocker.generateRawToken("resourceId", "userIdentity", 12 * 60);
-        CommunicationTokenCredential tokenCredential = new CommunicationTokenCredential(immediateFresher, tokenStr, true);
+        TokenRefreshOptions tokenRefreshOptions = new TokenRefreshOptions(immediateFresher, true);
+        CommunicationTokenCredential tokenCredential = new CommunicationTokenCredential(tokenRefreshOptions, tokenStr);
         assertTrue(tokenCredential.hasProactiveFetcher());
         tokenCredential.close();
         assertFalse(tokenCredential.hasProactiveFetcher());
@@ -262,7 +267,8 @@ public class CommunicationTokenCredentialTests {
     @Test
     public void shouldNotModifyTokenWhenRefresherThrows() throws InterruptedException, ExecutionException, IOException {
         String tokenStr = tokenMocker.generateRawToken("resourceId", "userIdentity", 601);
-        CommunicationTokenCredential tokenCredential = new CommunicationTokenCredential(exceptionRefresher, tokenStr, true);
+        TokenRefreshOptions tokenRefreshOptions = new TokenRefreshOptions(exceptionRefresher, true);
+        CommunicationTokenCredential tokenCredential = new CommunicationTokenCredential(tokenRefreshOptions, tokenStr);
         CountDownLatch countDownLatch = new CountDownLatch(1);
         exceptionRefresher.setOnCallReturn(countDownLatch::countDown);
 
@@ -277,7 +283,8 @@ public class CommunicationTokenCredentialTests {
     public void doNotSwallowExceptionWithoutProactiveFetching() throws InterruptedException, ExecutionException, IOException {
         String tokenStr = tokenMocker.generateRawToken("resourceId", "userIdentity", -5 * 60);
         exceptionRefresher.resetCallCount();
-        CommunicationTokenCredential tokenCredential = new CommunicationTokenCredential(exceptionRefresher, tokenStr, false);
+        TokenRefreshOptions tokenRefreshOptions = new TokenRefreshOptions(exceptionRefresher, false);
+        CommunicationTokenCredential tokenCredential = new CommunicationTokenCredential(tokenRefreshOptions, tokenStr);
         assertThrows(Exception.class, () -> {
             tokenCredential.getToken();
         }, "Should not swallow exception when client throws");
@@ -359,7 +366,8 @@ public class CommunicationTokenCredentialTests {
         longRunningRefresher.setOnCallReturn(countDownLatch::countDown);
         String tokenStr = tokenMocker.generateRawToken("resourceId", "userIdentity", -5 * 60);
 
-        CommunicationTokenCredential tokenCredential = new CommunicationTokenCredential(longRunningRefresher, tokenStr, true);
+        TokenRefreshOptions tokenRefreshOptions = new TokenRefreshOptions(longRunningRefresher, true);
+        CommunicationTokenCredential tokenCredential = new CommunicationTokenCredential(tokenRefreshOptions, tokenStr);
 
         countDownLatch.await();
         assertEquals(1, longRunningRefresher.numCalls());
@@ -374,7 +382,8 @@ public class CommunicationTokenCredentialTests {
     public void withoutInitialTokenShouldCallFresherOnlyOnceWhileRefreshingIsInProgress()
         throws InterruptedException, ExecutionException, IOException {
         longRunningRefresher.resetCallCount();
-        CommunicationTokenCredential tokenCredential = new CommunicationTokenCredential(longRunningRefresher);
+        TokenRefreshOptions tokenRefreshOptions = new TokenRefreshOptions(longRunningRefresher, true);
+        CommunicationTokenCredential tokenCredential = new CommunicationTokenCredential(tokenRefreshOptions);
         tokenCredential.getToken();
         assertEquals(1, longRunningRefresher.numCalls());
         for (int i = 0; i < 3; i++) {
@@ -439,7 +448,8 @@ public class CommunicationTokenCredentialTests {
     public void shouldRefreshAgainAfterClientFutureThrows() throws InterruptedException, ExecutionException,
             IOException {
         exceptionFutureRefresher.resetCallCount();
-        CommunicationTokenCredential tokenCredential = new CommunicationTokenCredential(exceptionFutureRefresher);
+        TokenRefreshOptions tokenRefreshOptions = new TokenRefreshOptions(exceptionFutureRefresher, true);
+        CommunicationTokenCredential tokenCredential = new CommunicationTokenCredential(tokenRefreshOptions);
         assertEquals(0, exceptionFutureRefresher.numCalls());
         Future<AccessToken> accessToken = tokenCredential.getToken();
         assertEquals(1, exceptionFutureRefresher.numCalls());
@@ -461,7 +471,8 @@ public class CommunicationTokenCredentialTests {
 
     @Test
     public void shouldThrowWhenGetTokenCalledOnClosedObject() throws IOException {
-        CommunicationTokenCredential tokenCredential = new CommunicationTokenCredential(exceptionFutureRefresher);
+        TokenRefreshOptions tokenRefreshOptions = new TokenRefreshOptions(exceptionFutureRefresher, true);
+        CommunicationTokenCredential tokenCredential = new CommunicationTokenCredential(tokenRefreshOptions);
         tokenCredential.close();
         assertThrows(RuntimeException.class, () -> {
             tokenCredential.getToken();