+++ sdk/communication/azure-communication-common/CHANGELOG.md
@@ -3,6 +3,8 @@
 ## 1.0.0-beta.4 (Unreleased)
 ### Breaking Changes
 - Renamed `CommunicationUserCredential` to `CommunicationTokenCredential`
+- Replaced constructor `CommunicationTokenCredential(TokenRefresher tokenRefresher, String initialToken,	boolean refreshProactively)` with `CommunicationTokenCredential(TokenRefreshOptions tokenRefreshOptions, String initialToken)`
+- Replaced constructor `CommunicationTokenCredential(TokenRefresher tokenRefresher)` with `CommunicationTokenCredential(TokenRefreshOptions tokenRefreshOptions)`
 
 ## 1.0.0-beta.3 (2020-11-16)
 Updated `azure-communication-common` version